backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.27.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.9
